# up: Use branch graph to avoid reloading state